% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy}
\usage{
accuracy(observed, predicted, na.rm = TRUE)
}
\arguments{
\item{observed}{(\code{factor}) The observed values. It has to have the same
length as \code{predicted}.}

\item{predicted}{(\code{factor}) The observed values. It has to have the same
length as \code{observed}.}

\item{na.rm}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
A single numeric value with the accuracy.
}
\description{
Given the observed and predicted values of categorical data (of any number of
classes) computes the accuracy (also known as proportion of correctly
classified cases).
}
\examples{
\dontrun{
accuracy(c("a", "b"), c("a", "b"))
accuracy(c("a", "b"), c("b", "a"))
accuracy(c("a", "b"), c("b", "b"))
accuracy(c("a", "b", "a"), c("b", "a", "c"))
}

}
\seealso{
Other categorical_metrics: 
\code{\link{brier_score}()},
\code{\link{categorical_summary}()},
\code{\link{confusion_matrix}()},
\code{\link{f1_score}()},
\code{\link{kappa_coeff}()},
\code{\link{matthews_coeff}()},
\code{\link{pccc}()},
\code{\link{pcic}()},
\code{\link{pr_auc}()},
\code{\link{precision}()},
\code{\link{recall}()},
\code{\link{roc_auc}()},
\code{\link{sensitivity}()},
\code{\link{specificity}()}
}
\concept{categorical_metrics}
