% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{best_lines_match}
\alias{best_lines_match}
\title{Percentage of best lines present in the predictions}
\usage{
best_lines_match(predictions, percentage = 10)
}
\arguments{
\item{predictions}{(\code{data.frame}) The predictions to evaluate. It must have
the columns \code{Line}, \code{Observed} and \code{Predicted}.}

\item{percentage}{(\code{numeric}) The percentage of best lines to consider. 10 by
default.}
}
\value{
A single numeric value with the percentage of best lines present in the
predictions.
}
\description{
Computes the percentage of best lines present in the predictions. The lines
with the highest observed values are considered as the best lines, and then,
the lines with the highest predicted values are compared to compute the
percentage of matching.
}
\examples{
\dontrun{


}

}
\seealso{
Other rank_metrics: 
\code{\link{ndcg}()}
}
\concept{rank_metrics}
