% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cholesky}
\alias{cholesky}
\title{Compute Cholesky}
\usage{
cholesky(x)
}
\arguments{
\item{x}{(\code{matrix}) The data to compute the cholesky.}
}
\value{
The cholesky \code{matrix}.
}
\description{
Compute the Choleski factorization of a real symmetric positive-definite
square matrix. If it fails, compute the Cholesky factorization of a non-real
symmetric positive-definite square matrix
}
