% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cholesky_no_definite}
\alias{cholesky_no_definite}
\title{Cholesky}
\usage{
cholesky_no_definite(G, tolerance = 1e-10)
}
\arguments{
\item{G}{(\code{numeric - matrix}) an object to apply this method, it could
be non positive-definite matrices.}

\item{tolerance}{(\code{double}) Tolerance level, by default is 1e-10.}
}
\description{
Compute the Cholesky factorization of a non-real symmetric
positive-definite square matrix. (Taken from BMTME R package at
https://github.com/frahik/BMTME/blob/master/R/cholesky.R).
}
