% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{class_mode}
\alias{class_mode}
\title{Get the value that appears more times. If there are more than 1
value that appears more times (multimodal), return the lowest value.}
\usage{
class_mode(x, na.rm = TRUE)
}
\description{
Get the value that appears more times. If there are more than 1
value that appears more times (multimodal), return the lowest value.
}
