% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{coef.Model}
\alias{coef.Model}
\title{Model coefficients}
\usage{
\method{coef}{Model}(model)
}
\arguments{
\item{model}{(\code{Model}) An object of a fitted model.}
}
\value{
\subsection{For univariate models}{

When the response variable is a numeric response, a named vector is returned
where each element maps to an element of the predictors matrix. Some models'
coefficients includes an intercept too.

When the response variable is categorical, including binary, a numeric matrix
is returned where the columns maps to the columns of the predictors matrix
and the rows corresponds to the different classes. Some models' coefficients
includes an additional row with the overall coefficient.
}

\subsection{For multivariate models}{

A named \code{list} with an element for each response variable in the fitted
model. Each element contains a vector with the coefficients of each response.
}
}
\description{
Extracts the model coefficients from a fitted model object.
}
\details{
The only models you can extract the coefficients are:
\code{\link[=generalized_linear_model]{generalized_linear_model()}}, \code{\link[=random_forest]{random_forest()}}, \code{\link[=partial_least_squares]{partial_least_squares()}}
and \code{\link[=bayesian_model]{bayesian_model()}}.
}
\examples{
\dontrun{
# Univariate analysis -------------------------------------------------------
x <- to_matrix(iris[, -5])
y <- iris$Species
model <- random_forest(x, y)

# Obtain the variables importance
coef(model)

# Multivariate analysis -----------------------------------------------------
x <- to_matrix(iris[, -c(1, 2)])
y <- iris[, c(1, 2)]
model <- generalized_linear_model(x, y)

# Obtain the models' coefficients
coef(model)
}

}
