% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{col_max_idx}
\alias{col_max_idx}
\title{col_max_idx: colvec max value index}
\usage{
col_max_idx(u, wlmt)
}
\arguments{
\item{wlmt:}{a limit search on colvec on indices within wlmt}
}
\value{
return an index of max value w.r.t to original index
}
\description{
col_max_idx: colvec max value index
}

