% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion matrix}
\usage{
confusion_matrix(observed, predicted, remove_na = TRUE)
}
\arguments{
\item{observed}{(\code{factor}) The observed values. It has to have the same
length as \code{predicted}.}

\item{predicted}{(\code{factor}) The observed values. It has to have the same
length as \code{observed}.}

\item{remove_na}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
An object of class \code{table} with the confusion matrix.
}
\description{
Given the observed and predicted values of categorical data (of any number of
classes) computes the confusion matrix.
}
\examples{
\dontrun{
confusion_matrix(factor("a"), factor("a"))
confusion_matrix(factor("a"), factor("b"))
confusion_matrix(factor(c("a", "b")), factor(c("c", "d")))
confusion_matrix(factor(c("a", "a")), factor(c("a", "a")))
confusion_matrix(iris$Species, iris$Species)
}

}
\seealso{
Other categorical_metrics: 
\code{\link{accuracy}()},
\code{\link{brier_score}()},
\code{\link{categorical_summary}()},
\code{\link{f1_score}()},
\code{\link{kappa_coeff}()},
\code{\link{matthews_coeff}()},
\code{\link{mode}()},
\code{\link{pccc}()},
\code{\link{pcic}()},
\code{\link{pr_auc}()},
\code{\link{precision}()},
\code{\link{recall}()},
\code{\link{roc_auc}()},
\code{\link{sensitivity}()},
\code{\link{specificity}()}
}
\concept{categorical_metrics}
