% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_kfold}
\alias{cv_kfold}
\title{K-fold cross validation folds generation}
\usage{
cv_kfold(records_number, k = 5)
}
\arguments{
\item{records_number}{(\code{numeric(1)}) The expected number of elements to be
included in the folds.}

\item{k}{(\code{numeric(1)}) The number of folds. 5 by default.}
}
\value{
A \code{list} with \code{k} elements where each element is a named \code{list} with the
elements \code{training} wich includes the indices of those records to be part of
the training set and \code{testing} wich includes the indices of those records to
be part of the testing set. Training and testing sets of each fold are
exhaustive and mutually exclusive.
}
\description{
Generates folds for the classic k-fold cross validation where k mutually
exclusive folds are generated and the training phase is done using k − 1
folds and the testing with the remaining one, which ensures all individuals
are part of the testing once.
}
\examples{
\dontrun{
# Generates 5 folds of 2 elements (10 / 5) in testing set
folds <- cv_kfold(10, 5)
# Indices of training set in fold 1
folds[[1]]$training
# Indices of testing set in fold 1
folds[[1]]$testing
folds[[2]]$training
folds[[2]]$testing

folds <- cv_kfold(100, 30)
# List with indices of training and testing of fold 1
folds[[1]]
# List with indices of training and testing of fold 2
folds[[2]]
folds[[3]]
# ...
folds[[30]]
}

}
