% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_leve_one_group_out}
\alias{cv_leve_one_group_out}
\title{Leave one group out cross validation folds generation}
\usage{
cv_leve_one_group_out(data)
}
\arguments{
\item{data}{(\code{vector}) The categorical data considered to stratify the folds.}
}
\value{
A \code{list} with \code{length(unique(data))} elements where each element is a named
\code{list} with the elements \code{training} wich includes the indices of those
records to be part of the training set and \code{testing} wich includes the
indices of those records to be part of the testing set. Training and testing
sets of each fold are exhaustive and mutually exclusive.
}
\description{
Generates folds for the leave one group out cross validation. In each fold
on group is taken to be the testing set and all the remaining ones to be the
training set.
}
\examples{
\dontrun{
# Generates 5 folds of 2 elements (10 / 5) in testing set
data <- c(rep("A", 10), rep("B", 20), rep("C", 30))
folds <- cv_leve_one_group_out(data)
# Indices of training set in fold 1
folds[[1]]$training
# Indices of testing set in fold 1
folds[[1]]$testing

# Verify fold 1 only contains elements of group A
table(data[folds[[1]]$training])
table(data[folds[[1]]$testing])
# Verify fold 2 only contains elements of group B
table(data[folds[[2]]$training])
table(data[folds[[2]]$testing])
# Verify fold 3 only contains elements of group C
table(data[folds[[3]]$training])
table(data[folds[[3]]$testing])
}

}
