% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_na}
\alias{cv_na}
\title{NA cross validation folds generation}
\usage{
cv_na(x)
}
\arguments{
\item{x}{(\code{vector}) The vector of values.}
}
\value{
A \code{list} with a single element which is a named \code{list} with the elements
\code{training} wich includes the indices of those records to be part of the
training set and \code{testing} wich includes the indices of those records to be
part of the testing set.
}
\description{
This type of cross validation consists of a single fold where all the NA
values are taken to be the testing set and the remaining ones to be the
training set. Useful in genomic selection when we want to predict the
phenotypes of the individuals with missing phenotypes.
}
\examples{
\dontrun{
# Generates random data with NAs
x <- c(1, 2, 3, NA, 5, 6, 7, NA, 9, 10)
folds <- cv_na(x)
# Indices of training set in fold 1
folds[[1]]$training
# Indices of testing set in fold 1
folds[[1]]$testing
}

}
