% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_random_strata}
\alias{cv_random_strata}
\title{Stratified random cross validation folds generation}
\usage{
cv_random_strata(data, folds_number = 5, testing_proportion = 0.2)
}
\arguments{
\item{data}{(\code{vector}) The categorical data considered to stratify the folds.}

\item{folds_number}{(\code{numeric(1)}) The number of folds. 5 by default.}

\item{testing_proportion}{(\code{numeric(1)}) The proportion of elements to be
included in the testing set in each fold. 0.2 by default.}
}
\value{
A \code{list} with \code{folds_number} elements where each element is a named \code{list}
with the elements \code{training} wich includes the indices of those records to be
part of the training set and \code{testing} wich includes the indices of those
records to be part of the testing set. Training and testing sets of each fold
are exhaustive and mutually exclusive.
}
\description{
Generates folds for the stratified random cross validation where you specify
the number of folds and the proportion of testing. In each fold a sample
without replacement of the specified proportion of testing individuals is
taken to be the testing set and all the remaining ones to be the training
set, but ensuring each fold contains the same proportion of elements in data.
}
\examples{
\dontrun{
# Generates 5 folds of 2 elements (10 / 5) in testing set
data <- c(rep("A", 10), rep("B", 20), rep("C", 30))
folds <- cv_random_strata(data, 5, 0.2)
# Indices of training set in fold 1
folds[[1]]$training
# Indices of testing set in fold 1
folds[[1]]$testing
# Verify fold 1 is balanced in training
table(data[folds[[1]]$training])
# Verify fold 1 is balanced in testing
table(data[folds[[1]]$testing])
#' # Verify fold 2 is balanced in training
table(data[folds[[2]]$training])
# Verify fold 2 is balanced in testing
table(data[folds[[2]]$testing])

folds <- cv_random_strata(iris$Species, 10, 0.5)
# List with indices of training and testing of fold 1
folds[[1]]
# List with indices of training and testing of fold 2
folds[[2]]
folds[[3]]
# ...
folds[[30]]
}

}
