% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skm.r
\name{dist_wlatlng}
\alias{dist_wlatlng}
\title{dist_wlatlng: dist btwn coordinate1<lat1, lng1> and coordinate2<lat2, lng2>}
\usage{
dist_wlatlng(.lat1, .lng1, .lat2, .lng2, .measure = "mi")
}
\arguments{
\item{.measure}{- mi or km}
}
\description{
dist_wlatlng: dist btwn coordinate1<lat1, lng1> and coordinate2<lat2, lng2>
}

