% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_wlatlng_km_cpp}
\alias{dist_wlatlng_km_cpp}
\title{dist_wlatlng_mi_cpp: dist between coordinate1<lat1, lng1> and coordinate2<lat2, lng2> in kilometer}
\usage{
dist_wlatlng_km_cpp(lat1, lng1, lat2, lng2)
}
\description{
dist_wlatlng_mi_cpp: dist between coordinate1<lat1, lng1> and coordinate2<lat2, lng2> in kilometer
}

