% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{echo}
\alias{echo}
\title{Print a message in the console.}
\usage{
echo(format, ..., end = "\\n", file = "", append = TRUE)
}
\arguments{
\item{format}{(\code{character(1)}) The format of the message. See \code{sprintf} for
more details.}

\item{...}{(\code{any}) The values to be formatted.}

\item{end}{(\code{character(1)}) The end of the message. Default is \verb{\\n}.}

\item{file}{(\code{character(1)}) The file where the output will be written.
Default is \code{""}, no file.}

\item{append}{(\code{logical(1)}) If \code{TRUE}, append the message to the file.}
}
\description{
This function is a wrapper of \code{cat} function, but it is more flexible
because it allows to use \code{sprintf} format.
}
\examples{
\dontrun{
echo("Hello \%s!", "World")
echo("\%d + \%d = \%d", 1, 1, 2)
}

}
