% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{f1_score}
\alias{f1_score}
\title{F1 score}
\usage{
f1_score(observed, predicted, positive_class = NULL, remove_na = TRUE)
}
\arguments{
\item{observed}{(\code{factor}) The observed values. It has to have the same
length as \code{predicted}.}

\item{predicted}{(\code{factor}) The observed values. It has to have the same
length as \code{observed}.}

\item{positive_class}{(\code{character(1)}) The name of the class (level) to be
taken as reference as the positive class. This parameter is only used for
binary variables. \code{NULL} by default which uses the second class in the
union of the classes (levels) in \code{observed} and \code{predicted}.}

\item{remove_na}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
For binary data a single value is returned, for more than 2 categories a
vector of F1 scores is returned, one per each category.
}
\description{
Given the observed and predicted values of categorical data (of any number of
classes) computes the F1 score, that combines the \link{precision} and \link{recall},
and it is defined as the harmonic mean of the precision and recall.
}
\details{
F1 score is computed as:\preformatted{![](f1_score.png "2 * ((precision * recall) / (precision + recall))")
}

See \code{\link[=precision]{precision()}} and \code{\link[=recall]{recall()}} for more information.
}
\examples{
\dontrun{
f1_score(factor(c("a", "b")), factor(c("a", "b")))
f1_score(factor(c("a", "b", "a", "b")), factor(c("a", "b", "b", "a")))
f1_score(factor(c("a", "b")), factor(c("b", "b")))
}

}
\seealso{
Other categorical_metrics: 
\code{\link{accuracy}()},
\code{\link{brier_score}()},
\code{\link{categorical_summary}()},
\code{\link{confusion_matrix}()},
\code{\link{kappa_coeff}()},
\code{\link{matthews_coeff}()},
\code{\link{mode}()},
\code{\link{pccc}()},
\code{\link{pcic}()},
\code{\link{pr_auc}()},
\code{\link{precision}()},
\code{\link{recall}()},
\code{\link{roc_auc}()},
\code{\link{sensitivity}()},
\code{\link{specificity}()}
}
\concept{categorical_metrics}
