% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_fast_bayesian.R
\name{gs_fast_bayesian}
\alias{gs_fast_bayesian}
\title{Fast Bayesian Cross Validation}
\usage{
gs_fast_bayesian(
  Pheno,
  Geno,
  traits,
  folds,
  is_multitrait = FALSE,
  iterations_number = 1500,
  burn_in = 500,
  thinning = 5,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{Pheno}{(\code{data.frame}) The phenotypic data. \code{Env} and \code{Line} columns
are required.}

\item{Geno}{(\code{matrix}) The genotypic data. It must be a square matrix with
the row and column names set to the line names in \code{Pheno}.}

\item{traits}{(\code{character}) The columns' names in \code{Pheno} to be used as
traits.}

\item{folds}{(\code{list}) A list of folds. Each fold is a named list with two
entries: \code{training}, with a vector of indices for training set, and
\code{testing}, with a vector of indices for testing set. Note that this is
default format for \verb{cv_*} functions of SKM libraries.}

\item{is_multitrait}{(\code{logical(1)}) Is multitrait analysis? \code{FALSE} by
default.}

\item{iterations_number}{(\code{numeric(1)}) Number of iterations to fit the
model. 1500 by default.}

\item{burn_in}{(\code{numeric(1)}) Number of items to burn at the beginning of the
model. 500 by default.}

\item{thinning}{(\code{numeric(1)}) Number of items to thin the model. 5 by
default.}

\item{seed}{(\code{numeric(1)}) A value to be used as internal seed for
reproducible results. \code{NULL} by default.}

\item{verbose}{(\code{logical(1)}) Should the progress information be printed?
\code{TRUE} by default.}

\item{validate_params}{(\code{logical(1)}) Should the parameters be validated? It
is not recommended to set this parameter to \code{FALSE} because if something
fails a non meaningful error is going to be thrown. \code{TRUE} by default.}
}
\value{
A \code{GSFastBayesian} object with the following attributes:
\itemize{
\item Pheno: (\code{data.frame}) The phenotypic data.
\item Geno: (\code{matrix}) The genotypic data.
\item traits: (\code{character}) The traits' names.
\item is_multitrait: (\code{logical(1)}) Is multitrait analysis?
\item Predictions: (\code{data.frame}) The predictions of cross validation. This
\code{data.frame} contains the \code{Trait}, \code{Fold}, \code{Line}, \code{Env}, \code{Predicted}
and \code{Observed} columns.
\item execution_time: (\code{difftime}) The execution time taken for the analysis.
\item folds: (\code{list}) The folds used in the analysis.
\item model: (\code{BayesianModel}) The model fitted.
\item model_name: (\code{character(1)}) The name of the model.
\item iterations_number: (\code{numeric(1)}) Number of iterations to fit the model.
\item burn_in: (\code{numeric(1)}) Number of items to burn at the beginning of the
model.
\item thinning: (\code{numeric(1)}) Number of items to thin the model.
}
}
\description{
This function performs a cross validation using the Fast Bayesian method
which requires to fit only one model per trait.
}
\examples{
data(Maize)

folds <- cv_kfold(nrow(Maize$Pheno), k = 5)

model <- gs_fast_bayesian(
  Maize$Pheno,
  Maize$Geno,

  traits = "Y",
  folds = folds,

  is_multitrait = FALSE,

  iterations_number = 10,
  burn_in = 5,
  thinning = 5,

  seed = NULL,
  verbose = TRUE
)

model

model$Predictions
}
\concept{gs_models}
