% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hush}
\alias{hush}
\title{Hide code output}
\usage{
hush(code, all = FALSE)
}
\arguments{
\item{code}{(\code{function}) The code to be evaluated}

\item{all}{(\code{logical}) If \code{TRUE}, suppress also warnings and the
result of the code. \code{FALSE} by default.}
}
\value{
If \emph{all} is \code{FALSE}, return whatever the evaluated code
returns, nothing otherwise.
}
\description{
Hide code output
}
