% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_empty}
\alias{is_empty}
\title{Is an empty object.}
\usage{
is_empty(x)
}
\arguments{
\item{x}{(\code{any}) The object to be checked.}
}
\value{
A logical value indicating if the object is empty.
}
\description{
Check if an object is empty.
}
\examples{
\dontrun{
is_empty(1:10) # FALSE
is_empty(1:0) # TRUE
is_empty(NULL) # TRUE
is_empty(c()) # TRUE
is_empty(character()) # TRUE
is_empty(data.frame()) # TRUE
}

}
