% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{maape}
\alias{maape}
\title{Mean Arctangent Absolute Percentage Error}
\usage{
maape(observed, predicted, remove_na = TRUE)
}
\arguments{
\item{observed}{(\code{numeric}) The observed values. It has to have the same
length as \code{predicted}.}

\item{predicted}{(\code{numeric}) The observed values. It has to have the same
length as \code{observed}.}

\item{remove_na}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
A single numeric value with the Mean Arctangent Absolute Percentage Error.
}
\description{
Given the observed and predicted values of numeric data computes the
Mean Arctangent Absolute Percentage Error.
}
\details{
Mean Arctangent Absolute Percentage Error is computed as:\preformatted{![](maape.png "mean(atan(abs(observed - predicted) / abs(observed)))")
}

where y_i is the observed value of element i, hat{y_i} is the predicted
value of element i and N is the total number of elements.
}
\examples{
\dontrun{
set.seed(1)
x <- rnorm(100)
maape(x, x)
maape(x, x - 1)
maape(x, x + 10)
maape(x, x + 10)
maape(x, x + 10)
}

}
\seealso{
Other numeric_metrics: 
\code{\link{mae}()},
\code{\link{mse}()},
\code{\link{nrmse}()},
\code{\link{numeric_summary}()},
\code{\link{pearson}()},
\code{\link{rmse}()},
\code{\link{spearman}()}
}
\concept{numeric_metrics}
