% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mae}
\alias{mae}
\title{Mean Abosolute Error}
\usage{
mae(observed, predicted, remove_na = TRUE)
}
\arguments{
\item{observed}{(\code{numeric}) The observed values. It has to have the same
length as \code{predicted}.}

\item{predicted}{(\code{numeric}) The observed values. It has to have the same
length as \code{observed}.}

\item{remove_na}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
A single numeric value with the Mean Abosolute Error.
}
\description{
Given the observed and predicted values of numeric data computes the
Mean Abosolute Error.
}
\examples{
\dontrun{
set.seed(1)
x <- rnorm(100)
mae(x, x)
mae(x, x - 1)
mae(x, x + 10)
mae(x, x + 10)
mae(x, x + 10)
}

}
\seealso{
Other numeric_metrics: 
\code{\link{maape}()},
\code{\link{mse}()},
\code{\link{nrmse}()},
\code{\link{numeric_summary}()},
\code{\link{pearson}()},
\code{\link{rmse}()},
\code{\link{spearman}()}
}
\concept{numeric_metrics}
