% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Maize}
\alias{Maize}
\title{Genomic Maize data.}
\format{
A list with two fields: Pheno and Geno. Pheno is a data frame with
2888 observations and 4 variables. Geno is a data frame with 722 observations
of 723 variables.
}
\source{
M. Bandeira e Sousa et al., "Genomic-Enabled Prediction in Maize
Using Kernel Models with Genotype x Environment Interaction"
\url{https://doi.org/10.1534/g3.117.042341}
}
\usage{
Maize
}
\description{
Phenotypic and Genotype maize data.
}
\keyword{datasets}
