% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{matthews_coeff}
\alias{matthews_coeff}
\title{Matthews Correlation Coefficient (MCC)}
\usage{
matthews_coeff(observed, predicted, remove_na = TRUE)
}
\arguments{
\item{observed}{(\code{factor}) The observed values. It has to have the same
length as \code{predicted}.}

\item{predicted}{(\code{factor}) The observed values. It has to have the same
length as \code{observed}.}

\item{remove_na}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
A single numeric value with the Matthews Correlation Coefficient.
}
\description{
Given the observed and predicted values of binary data computes the Matthews
Correlation Coefficient (MCC) also known as Phi Coefficient or Mean Square
Contingency Coefficient.
}
\examples{
\dontrun{
matthews_coeff(factor(c("a", "b")), factor(c("a", "b")))
matthews_coeff(factor(c("a", "b")), factor(c("b", "a")))
matthews_coeff(factor(c("a", "b")), factor(c("b", "b")))
matthews_coeff(factor(c(TRUE, FALSE)), factor(c(FALSE, TRUE)))
}

}
\seealso{
Other categorical_metrics: 
\code{\link{accuracy}()},
\code{\link{brier_score}()},
\code{\link{categorical_summary}()},
\code{\link{confusion_matrix}()},
\code{\link{f1_score}()},
\code{\link{kappa_coeff}()},
\code{\link{mode}()},
\code{\link{pccc}()},
\code{\link{pcic}()},
\code{\link{pr_auc}()},
\code{\link{precision}()},
\code{\link{recall}()},
\code{\link{roc_auc}()},
\code{\link{sensitivity}()},
\code{\link{specificity}()}
}
\concept{categorical_metrics}
