% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mkdir}
\alias{mkdir}
\title{Make directory}
\usage{
mkdir(directory)
}
\arguments{
\item{directory}{(\code{character(1)}) The directory to create.}
}
\description{
This functions  creates a directory if it does not exists.
}
\examples{
\dontrun{
mkdir("my_new_directory")
}

}
