% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mse}
\alias{mse}
\title{Mean Squared Error}
\usage{
mse(observed, predicted, remove_na = TRUE)
}
\arguments{
\item{observed}{(\code{numeric}) The observed values. It has to have the same
length as \code{predicted}.}

\item{predicted}{(\code{numeric}) The observed values. It has to have the same
length as \code{observed}.}

\item{remove_na}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
A single numeric value with the Mean Squared Error.
}
\description{
Given the observed and predicted values of numeric data computes the Mean
Squared Error.
}
\details{
Mean Squared Error is computed as:\preformatted{![](mse.png "(1 / N) * sum((observed - predicted)^2)")
}

where y_i is the observed value of element i, hat{y_i} is the predicted
value of element i and N is the total number of elements.
}
\examples{
\dontrun{
mse(1, 1)
mse(1, 0)
mse(c(1, 2, NA, 4), c(1, NA, 3, 4))

set.seed(1)
x <- rnorm(100)
mse(x, x)
mse(x, x - 1)
mse(x, x + 10)
}

}
\seealso{
Other numeric_metrics: 
\code{\link{maape}()},
\code{\link{mae}()},
\code{\link{nrmse}()},
\code{\link{numeric_summary}()},
\code{\link{pearson}()},
\code{\link{rmse}()},
\code{\link{spearman}()}
}
\concept{numeric_metrics}
