% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{nrmse}
\alias{nrmse}
\title{Normalized Root Mean Squared Error}
\usage{
nrmse(observed, predicted, type = "sd", na.rm = TRUE)
}
\arguments{
\item{observed}{(\code{numeric}) The observed values. It has to have the same
length as \code{predicted}.}

\item{predicted}{(\code{numeric}) The observed values. It has to have the same
length as \code{observed}.}

\item{type}{(\code{character(1)}) (case not sensitive) The normalization type to
use. The options are \code{"sd"}, \code{"mean"}, \code{"maxmin"} (or \code{"range"}) and
\code{"iqr"} (for more information, see Details section below). \code{"sd"} by
default.}

\item{na.rm}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
A single numeric value with the Normalized Root Mean Squared Error.
}
\description{
Given the observed and predicted values of numeric data computes the
Normalized Root Mean Squared Error.
}
\details{
The formula is the same as \code{\link[=rmse]{rmse()}} (Root Mean Square Error) but divided by a
normalization term specified in \code{type}:
\itemize{
\item \code{"sd"}: Standard deviatione.
\item \code{"mean"}: Mean.
\item \code{"maxmin"} or \code{"range"}: The range, maximun minus minimum.
\item \code{"iqr"}: Interquantile range (Q3 - Q1).
}
}
\examples{
\dontrun{
set.seed(1)
x <- rnorm(100)
nrmse(x, x)
nrmse(x, x - 1, type = "mean")
nrmse(x, x + 10, type = "iqr")
nrmse(x, x + 10, type = "range")
nrmse(x, x + 10, type = "maxmin")
}

}
\seealso{
Other numeric_metrics: 
\code{\link{maape}()},
\code{\link{mae}()},
\code{\link{mse}()},
\code{\link{numeric_summary}()},
\code{\link{pearson}()},
\code{\link{rmse}()},
\code{\link{spearman}()}
}
\concept{numeric_metrics}
