% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{numeric_summary}
\alias{numeric_summary}
\title{Numeric summary}
\usage{
numeric_summary(observed, predicted, na.rm = TRUE)
}
\arguments{
\item{observed}{(\code{numeric}) The observed values. It has to have the same
length as \code{predicted}.}

\item{predicted}{(\code{numeric}) The observed values. It has to have the same
length as \code{observed}.}

\item{na.rm}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
A list with all metrics.
}
\description{
Given the observed and predicted values of numeric data computes the Mean
Squared Error (MSE), Root Mean Squared Error (RMSE), Normalized Root Mean
Squared Error (NRMSE), Mean Arctangent Absolute Percentage Error (MAAPE),
Mean Absolute Error (MAE) and Person's correlation
}
\examples{
\dontrun{
set.seed(1)
x <- rnorm(100)
numeric_summary(x, x)
numeric_summary(x, x - 1)
numeric_summary(x, x + 10)
}

}
\seealso{
Other numeric_metrics: 
\code{\link{maape}()},
\code{\link{mae}()},
\code{\link{mse}()},
\code{\link{nrmse}()},
\code{\link{pearson}()},
\code{\link{rmse}()},
\code{\link{spearman}()}
}
\concept{numeric_metrics}
