% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pearson}
\alias{pearson}
\title{Pearson's correlation}
\usage{
pearson(x, y = x, na.rm = TRUE)
}
\arguments{
\item{x}{(\code{numeric} | \code{matrix}) The values to calculate the correlation.}

\item{y}{(\code{numeric} | \code{matrix}) The values to calculate the correlation with.
the same values as \code{x} by default.}

\item{na.rm}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
A single numeric value with the Pearson's correlation.
}
\description{
Computes the Pearson's correlation.
}
\examples{
\dontrun{
set.seed(1)
x <- rnorm(100)
pearson(x, x)
pearson(x, x + rnorm(100, 2))
pearson(x, x - 1)
pearson(x, x + 10)
}

}
\seealso{
Other numeric_metrics: 
\code{\link{maape}()},
\code{\link{mae}()},
\code{\link{mse}()},
\code{\link{nrmse}()},
\code{\link{numeric_summary}()},
\code{\link{rmse}()},
\code{\link{spearman}()}
}
\concept{numeric_metrics}
