% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pr_auc}
\alias{pr_auc}
\title{Precision-Recall Area Under the Curver (PR-AUC)}
\usage{
pr_auc(observed, probabilities, positive_class = NULL, remove_na = TRUE)
}
\arguments{
\item{observed}{(\code{factor}) The observed values. It has to have the same
length as rows \code{probabilities}.}

\item{probabilities}{(\code{data.frame}) The probability of each class for each
individual. It is required the columns names of \code{probabilities} corresponds
to all classes (levels) in \code{observed} and that \code{probabilities} has as many
rows as records \code{observed}.}

\item{positive_class}{(\code{character(1)}) The name of the class (level) to be
taken as reference as the positive class. This parameter is only used for
binary variables. \code{NULL} by default which uses the second class in the
union of the classes (levels) in \code{observed} and \code{predicted}.}

\item{remove_na}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
A single numeric value with the PR-AUC.
}
\description{
Given the observed values and predicted probabilities values of binary data
computes the Precision-Recall Area Under the Curve (PR-AUC).
}
\examples{
\dontrun{
pr_auc(factor(c("a", "b")), data.frame(a = c(0.2, 0.6), b = c(0.5, 0.4)))
pr_auc(factor(c("a", "b")), data.frame(a = c(0.8, 0.3), b = c(0.2, 0.7)))
pr_auc(
  factor(c("a", "b")),
  data.frame(a = c(0.2, 0.6), b = c(0.5, 0.4)),
  positive_class = "b"
)
pr_auc(
  factor(c(TRUE, FALSE)),
  data.frame(
    `TRUE` = c(0.3, 0.2),
    `FALSE` = c(0.7, 0.8),
    check.names = FALSE
  )
)
}

}
\seealso{
Other categorical_metrics: 
\code{\link{accuracy}()},
\code{\link{brier_score}()},
\code{\link{categorical_summary}()},
\code{\link{confusion_matrix}()},
\code{\link{f1_score}()},
\code{\link{kappa_coeff}()},
\code{\link{math_mode}()},
\code{\link{matthews_coeff}()},
\code{\link{pccc}()},
\code{\link{pcic}()},
\code{\link{precision}()},
\code{\link{recall}()},
\code{\link{roc_auc}()},
\code{\link{sensitivity}()},
\code{\link{specificity}()}
}
\concept{categorical_metrics}
