% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error}
\usage{
rmse(observed, predicted, na.rm = TRUE)
}
\arguments{
\item{observed}{(\code{numeric}) The observed values. It has to have the same
length as \code{predicted}.}

\item{predicted}{(\code{numeric}) The observed values. It has to have the same
length as \code{observed}.}

\item{na.rm}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
A single numeric value with the Root Mean Squared Error.
}
\description{
Given the observed and predicted values of numeric data computes the Root
Mean Squared Error.
}
\examples{
\dontrun{
rmse(1, 1)
rmse(1, 0)
rmse(c(1, 2, NA, 4), c(1, NA, 3, 4))

set.seed(1)
x <- rnorm(100)
rmse(x, x)
rmse(x, x - 1)
rmse(x, x + 10)
}

}
\seealso{
Other numeric_metrics: 
\code{\link{maape}()},
\code{\link{mae}()},
\code{\link{mse}()},
\code{\link{nrmse}()},
\code{\link{numeric_summary}()},
\code{\link{pearson}()},
\code{\link{spearman}()}
}
\concept{numeric_metrics}
