% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sensitivity}
\alias{sensitivity}
\title{Sensitivity}
\usage{
sensitivity(observed, predicted, positive_class = NULL, remove_na = TRUE)
}
\arguments{
\item{observed}{(\code{factor}) The observed values. It has to have the same
length as \code{predicted}.}

\item{predicted}{(\code{factor}) The observed values. It has to have the same
length as \code{observed}.}

\item{positive_class}{(\code{character(1)}) The name of the class (level) to be
taken as reference as the positive class. This parameter is only used for
binary variables. \code{NULL} by default which uses the second class in the
union of the classes (levels) in \code{observed} and \code{predicted}.}

\item{remove_na}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
For binary data a single value is returned, for more than 2 categories a
vector of sensitivities is returned, one per each category.
}
\description{
Given the observed and predicted values of categorical data (of any number of
classes) computes the sensitivity (also known as recall), the metric that
evaluates a models ability to predict true positives of each available
category.
}
\examples{
\dontrun{
sensitivity(factor(c("a", "b")), factor(c("a", "b")))
sensitivity(factor(c("a", "b")), factor(c("b", "a")))
sensitivity(factor(c("a", "b")), factor(c("b", "b")))
sensitivity(factor(c(TRUE, FALSE)), factor(c(FALSE, TRUE)))
sensitivity(factor(c("a", "b", "a")), factor(c("b", "a", "c")))
}

}
\seealso{
Other categorical_metrics: 
\code{\link{accuracy}()},
\code{\link{brier_score}()},
\code{\link{categorical_summary}()},
\code{\link{confusion_matrix}()},
\code{\link{f1_score}()},
\code{\link{kappa_coeff}()},
\code{\link{matthews_coeff}()},
\code{\link{mode}()},
\code{\link{pccc}()},
\code{\link{pcic}()},
\code{\link{pr_auc}()},
\code{\link{precision}()},
\code{\link{recall}()},
\code{\link{roc_auc}()},
\code{\link{specificity}()}
}
\concept{categorical_metrics}
