% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skm.r
\docType{data}
\name{skm}
\alias{skm}
\title{skm: selective k-means
selective kmeans solve the following problem:
assume a data.table of s - t - d(s, t) for all combinations of s and t,
choose k of s that minimizes sum(min(d(s, t))) with k such s and all t.}
\format{An object of class \code{C++Class} of length 1.}
\usage{
skm
}
\arguments{
\item{x:}{data.table with s - t - d(s, t): s<source> - t<target> - d<dist>,
where s<source> and t<target> must characters and d<distance> must numeric.
aware d<distance> is not necessary as an euclidean or any distance and even
necessary as symmetric - d(s, t) can be unequal to d(t, s) - view d as such
a measure of the cost of assigning one to the other!}

\item{k:}{number of centers}

\item{s_colname:}{s<source>}

\item{t_colname:}{t<target>}

\item{d_colname:}{d<distance> - view d as cost of assigning t into s.
also modify the input data or build in the algorithm can solve problem with
a different fixed cost on using each s as source - i prefer to moddify data
so that the algorithm is clean and clear - i will show a how to in vignette}

\item{w_colname:}{w<weighting> - optional: when none null optimizing toward
objective to minimize d = d * w such as weighted cost of assigning t into s}

\item{s_must:}{length <= k-1 s must in result: conditional optimizing.}

\item{s_init:}{initialize optimization with s_init<must have length k>
if both s_must s_init none null then all(s_must %in% s_init) must be TRUE.}

\item{max_it:}{max number of iterations can run for optimizing result.}
}
\value{
list(s, t):
s - data.table with s<source> - k<0, 1..k> when 0 it not as being selected
when 1 - k it is selected as the <ik>th center - ik as value i in column k
t - data.table with t<target> - k<1..k> where k is assigning of t to <ik>s
}
\description{
skm: selective k-means
selective kmeans solve the following problem:
assume a data.table of s - t - d(s, t) for all combinations of s and t,
choose k of s that minimizes sum(min(d(s, t))) with k such s and all t.
}
\keyword{datasets}

