% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skm.r
\name{skm_mlp}
\alias{skm_mlp}
\title{skm_mlp: skm with multiple runs parallel}
\usage{
skm_mlp(x, k = 1L, s_colname = "s", t_colname = "t", d_colname = "d",
  max_it = 10L, s_init = NULL, n_init = 10L, min_at = 10L,
  max_at = 50L, halt_at = 20L)
}
\description{
skm_mlp: skm with multiple runs parallel
}

