% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skm_cpp_rd.r
\name{skm_rgi_cpp}
\alias{skm_rgi_cpp}
\title{skm_rgi_cpp}
\usage{
skm_rgi_cpp(x, k, s_must, max_it)
}
\arguments{
\item{x}{an m x n matrix often m < n, as a convention index rows of x with s, and
cols of x with t so x(i, j) can be expressed as (s_i, t_j) equally.}

\item{k}{number of index to be selected from x row index start from 0.}

\item{s_must}{an index vector set should be selected before selecting other index.}

\item{max_it}{max number of iterations can run for optimizing result.
max number of iterations within a single initial run on optimal path.}
}
\value{
skmSolution
}
\description{
solve skm with single and random size k s_init
}
\details{
refer skm_sgl_cpp
}
\seealso{
Other skm: \code{\link{skm_mlp_cpp}},
  \code{\link{skm_mls_cpp}}, \code{\link{skm_rgs_cpp}},
  \code{\link{skm_sgl_cpp}}
}

