% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skm_cpp_rd.r
\name{skmRpl_mlp_cpp}
\alias{skmRpl_mlp_cpp}
\title{skmRpl_mlp_cpp}
\usage{
skmRpl_mlp_cpp(x, k, s_must, max_it, max_at, skmRpl_GS = 100L)
}
\arguments{
\item{x}{an m x n matrix often m < n, as a convention index rows of x with s, and
cols of x with t so x(i, j) can be expressed as (s_i, t_j) equally.}

\item{k}{number of index to be selected from x row index start from 0.}

\item{s_must}{an index vector set should be selected before selecting other index.}

\item{max_it}{max number of iterations can run for optimizing result.
max number of iterations within a single initial run on optimal path.}

\item{max_at}{max number of attempts or repeats on running for optimial results,
max number of random initialization for finding optimial results.}

\item{skmRpl_GS}{skmRpl_GS: RcppParallel grain size when run skmRpl_mlp_cpp}
}
\value{
skmSolution
 skmSolution present in r list
}
\description{
solve skm with multiple runs in parallel
}
\details{
refer skm_sgl_cpp
}

