% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skm_cpp_rd.r
\name{skmSolution}
\alias{skmSolution}
\alias{Rcpp_skmSolution}
\alias{Rcpp_skmSolution-class}
\title{skmSolution}
\format{An object of class \code{C++Class} of length 1.}
\usage{
skmSolution
}
\description{
class skmSolution, which often returned via skm solver implemented in cpp
}
\details{
an skmSolution instance has two member variable:

  o: objective sum(min(x.subview(i in s, all j), min over all i), sum over all j)

  s: selected index set of row index start from 0
}
\keyword{datasets}

