% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{spearman}
\alias{spearman}
\title{Spearman's correlation coefficient}
\usage{
spearman(x, y = x, remove_na = TRUE)
}
\arguments{
\item{x}{(\code{numeric} | \code{matrix}) The values to calculate the correlation.}

\item{y}{(\code{numeric} | \code{matrix}) The values to calculate the correlation with.
the same values as \code{x} by default.}

\item{remove_na}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
A single numeric value with the Spearman's correlation.
}
\description{
Computes the Spearman's correlation.
}
\details{
Spearman's correlation coefficient is computed as:\preformatted{![](spearman.png "cov(rank(x), rank(y)) / (sd(rank(x)) * sd(rank(y)))")
}

where R is the rank of a variable, cov the covariance of two variables and
sigma is the standard deviation.
}
\examples{
\dontrun{
set.seed(1)
x <- rnorm(100)
spearman(x, x)
spearman(x, x + rnorm(100, 2))
spearman(x, x - 1)
spearman(x, x + 10)
}

}
\seealso{
Other numeric_metrics: 
\code{\link{maape}()},
\code{\link{mae}()},
\code{\link{mse}()},
\code{\link{nrmse}()},
\code{\link{numeric_summary}()},
\code{\link{pearson}()},
\code{\link{rmse}()}
}
\concept{numeric_metrics}
