% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stratified_sampling}
\alias{stratified_sampling}
\title{stratified_sampling}
\usage{
stratified_sampling(v, k, g)
}
\arguments{
\item{v:}{candidate v.}

\item{k:}{selection k.}

\item{g:}{stratify structure g.}
}
\value{
s: select from v length k stratified by g.
}
\description{
select k elements from vector v w.r.t stratify structure group g.
 TODO - implementing via template so v is flexible as vec or uvec.
}

