% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_data_frame}
\alias{to_data_frame}
\title{Convert data to data.frame}
\usage{
to_data_frame(x, na.rm = FALSE)
}
\arguments{
\item{x}{(\code{any}) The object to be converted to \code{data.frame}.}

\item{na.rm}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{FALSE} by
default.}
}
\value{
A matrix.
}
\description{
Given an object convert it to \code{data.frame}.
}
\details{
The following rules are applied when converting the object to matrix
depending the object's type:
\itemize{
\item \code{numeric} vector: It is converted to a one column \code{data.frame}.
\item \code{character}, \code{logical} and \code{factor} vectors: All of these are considered
to be categorical variables and therefore a coerced to \code{factor} in a one
column \code{data.frame}.
\item \code{matrix}: If it is \code{character}, \code{logical} or \code{factor} are included in the
same way as described for vectors and only coerced to \code{data.frame} for
\code{numeric} matrices.
}

For all the columns a name is assigned if they has no one.
}
\examples{
\dontrun{
to_data_frame(iris)
to_data_frame(1:10)
to_data_frame(c("a", "b", "c"))
to_data_frame(data.frame(a = c("a", "b"), b = c("c", "d")))
}

}
