% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wheat}
\alias{Wheat}
\title{Genomic Wheat data.}
\format{
A list with two fields: Pheno and Geno. Pheno is a data frame with
2396 observations and 4 variables. Geno is a data frame with 599 observations
of 600 variables.
}
\source{
J. Crossa et al., "Prediction of Genetic Values of Quantitative
Traits in Plant Breeding Using Pedigree and Molecular Markers"
\url{https://doi.org/10.1534/genetics.110.118521}
}
\usage{
Wheat
}
\description{
Phenotypic and Genotype wheat data.
}
\keyword{datasets}
