% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_csv}
\alias{write_csv}
\title{Write a CSV}
\usage{
write_csv(data, file, quote = "auto", na = "NA", ...)
}
\arguments{
\item{data}{(\code{data.frame}) The data to save.}

\item{file}{(\code{character(1)}) The file to save the data.}

\item{quote}{(\code{character(1)}) The quote strategy to use. "auto" by default.
See \code{data.table::fwrite} for more information.}

\item{na}{(\code{character(1)}) The string to use for NA values. "NA" by default.}

\item{...}{(\code{...}) Other arguments to pass to \code{data.table::fwrite}.}
}
\description{
Wrapper for the \code{data.table::fwrite} function fot saving large \code{data.frame}s
quickly in a CSV file.
}
\examples{
\dontrun{
write_csv(iris, "iris.csv")
write_csv(data.frame(a = 1:10, b = 1:10), "my_file.csv")
}

}
