% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{zip2012}
\alias{zip2012}
\title{zip2012}
\format{A data table with 28844 rows and 9 variables:
\describe{
  \item{zip}{zip code, 5 digits zip code in U.S.}
  \item{lat}{latitude}
  \item{lng}{longitude}
  \item{pop}{population}
  \item{ink}{income}
  \item{city}{city}
  \item{state}{state}
  \item{p_pop}{percentage of population w.r.t total population}
  \item{p_ink}{percentage of income w.r.t total income}
}}
\source{
\url{http://federalgovernmentzipcodes.us/}
}
\usage{
zip2012
}
\description{
a zip code database with latitude, longitude, population and income.
}
\keyword{datasets}

