% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{col_rgn_val}
\alias{col_rgn_val}
\title{col_rgn_val: colvec range = max - min value within limited range}
\usage{
col_rgn_val(u, wlmt)
}
\description{
col_rgn_val: colvec range = max - min value within limited range
}

