% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_fast_bayesian.R
\name{gs_fast_bayesian}
\alias{gs_fast_bayesian}
\title{Fast Bayesian Cross Validation for Genomic Selection}
\usage{
gs_fast_bayesian(
  Pheno,
  Geno,
  traits,
  folds,
  predictors = c("Env", "Line", "EnvxLine"),
  is_multitrait = FALSE,
  iterations_number = 1500,
  burn_in = 500,
  thinning = 5,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{Pheno}{(\code{data.frame}) The phenotypic data. \code{Env} and \code{Line} columns
are required.}

\item{Geno}{(\code{matrix}) The genotypic data. It must be a square matrix with
the row and column names set to the line names in \code{Pheno}.}

\item{traits}{(\code{character}) The columns' names in \code{Pheno} to be used as
traits.}

\item{folds}{(\code{list}) A list of folds. Each fold is a named list with two
entries: \code{training}, with a vector of indices for training set, and
\code{testing}, with a vector of indices for testing set. Note that this is
default format for \verb{cv_*} functions of SKM libraries.}

\item{predictors}{(\code{character}) (case not sensitive) The predictors to be
used in the model. "Env" stans for the environment effect, "Line" stands
for the line effect and "EnvxLine" stands for the interaction between
environment and line. "Env" and "Line" are required.
\code{c("Env", "Line", "EnvxLine")} by default.}

\item{is_multitrait}{(\code{logical(1)}) Is multitrait analysis? \code{FALSE} by
default.}

\item{iterations_number}{(\code{numeric(1)}) Number of iterations to fit the
model. 1500 by default.}

\item{burn_in}{(\code{numeric(1)}) Number of items to burn at the beginning of the
model. 500 by default.}

\item{thinning}{(\code{numeric(1)}) Number of items to thin the model. 5 by
default.}

\item{seed}{(\code{numeric(1)}) A value to be used as internal seed for
reproducible results. \code{NULL} by default.}

\item{verbose}{(\code{logical(1)}) Should the progress information be printed?
\code{TRUE} by default.}

\item{validate_params}{(\code{logical(1)}) Should the parameters be validated? It
is not recommended to set this parameter to \code{FALSE} because if something
fails a non meaningful error is going to be thrown. \code{TRUE} by default.}
}
\value{
A \code{GSFastBayesian} object with the following attributes:
\itemize{
\item Pheno: (\code{data.frame}) The phenotypic data.
\item Geno: (\code{matrix}) The genotypic data.
\item traits: (\code{character}) The traits' names.
\item is_multitrait: (\code{logical(1)}) Is multitrait analysis?
\item Predictions: (\code{data.frame}) The predictions of cross validation. This
\code{data.frame} contains the \code{Trait}, \code{Fold}, \code{Line}, \code{Env}, \code{Predicted}
and \code{Observed} columns.
\item execution_time: (\code{difftime}) The execution time taken for the analysis.
\item folds: (\code{list}) The folds used in the analysis.
\item model: (\code{BayesianModel}) The model fitted.
\item model_name: (\code{character(1)}) The name of the model.
\item iterations_number: (\code{numeric(1)}) Number of iterations to fit the model.
\item burn_in: (\code{numeric(1)}) Number of items to burn at the beginning of the
model.
\item thinning: (\code{numeric(1)}) Number of items to thin the model.
}
}
\description{
This function performs a cross validation using the Fast Bayesian method
which requires to fit only one model per trait.
}
\examples{
data(Maize)

folds <- cv_kfold(nrow(Maize$Pheno), k = 5)

results <- gs_fast_bayesian(
  Maize$Pheno,
  Maize$Geno,

  traits = "Y",
  folds = folds,

  is_multitrait = FALSE,

  iterations_number = 10,
  burn_in = 5,
  thinning = 5,

  seed = NULL,
  verbose = TRUE
)

print(results)
#> $Predictions
#> # A tibble: 2,888 × 6
#>   Trait Fold  Line      Env   Observed Predicted
#>   <fct> <fct> <fct>     <fct>    <dbl>     <dbl>
#> 1 Y     1     L001_L003 1      -0.0274      7.11
#> 2 Y     1     L001_L012 1       0.574       3.89
#> 3 Y     1     L001_L023 1      -0.0636      3.95
#> 4 Y     1     L001_L033 1       0.215       2.05
#> 5 Y     1     L001_L034 1       0.299       2.70
#> # … with 2,883 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#>
#> $traits: Y
#>
#> $is_multitrait: FALSE
#>
#> $folds: 5
#>
#> $execution_time: 0.841673 mins
#>
#> $Pheno
#>         envs_num: 4
#>         lines_num: 722
#>         rows_num: 2888
#>
#> $Geno
#>         rows_num: 722
#>         cols_num: 722
#>
#> $model_name: BGBLUP
#>
#> $iterations_number: 10
#>
#> $burn_in: 5
#>
#> $thinning: 5
}
\concept{gs_models}
