% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{kappa_coeff}
\alias{kappa_coeff}
\title{Cohen's Kappa coefficient}
\usage{
kappa_coeff(observed, predicted, remove_na = TRUE)
}
\arguments{
\item{observed}{(\code{factor}) The observed values. It has to have the same
length as \code{predicted}.}

\item{predicted}{(\code{factor}) The observed values. It has to have the same
length as \code{observed}.}

\item{remove_na}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
A single numeric value with the Cohen's Kappa coefficient.
}
\description{
Given the observed and predicted values of categorical data (of any number of
classes) computes the Cohen's Kappa coefficient.
}
\examples{
\dontrun{
kappa_coeff(factor(c("a", "b")), factor(c("a", "b")))
kappa_coeff(factor(c("a", "b")), factor(c("b", "a")))
kappa_coeff(factor(c("a", "b", "a")), factor(c("b", "a", "c")))
}

}
\seealso{
Other categorical_metrics: 
\code{\link{accuracy}()},
\code{\link{brier_score}()},
\code{\link{categorical_summary}()},
\code{\link{confusion_matrix}()},
\code{\link{f1_score}()},
\code{\link{matthews_coeff}()},
\code{\link{mode}()},
\code{\link{pccc}()},
\code{\link{pcic}()},
\code{\link{pr_auc}()},
\code{\link{precision}()},
\code{\link{recall}()},
\code{\link{roc_auc}()},
\code{\link{sensitivity}()},
\code{\link{specificity}()}
}
\concept{categorical_metrics}
