% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{nrmse}
\alias{nrmse}
\title{Normalized Root Mean Squared Error}
\usage{
nrmse(observed, predicted, type = "sd", remove_na = TRUE)
}
\arguments{
\item{observed}{(\code{numeric}) The observed values. It has to have the same
length as \code{predicted}.}

\item{predicted}{(\code{numeric}) The observed values. It has to have the same
length as \code{observed}.}

\item{type}{(\code{character(1)}) (case not sensitive) The normalization type to
use. The options are \code{"sd"}, \code{"mean"}, \code{"maxmin"} (or \code{"range"}) and
\code{"iqr"} (for more information, see Details section below). \code{"sd"} by
default.}

\item{remove_na}{(\code{logical(1)}) Should \code{NA} values be removed?. \code{TRUE} by
default.}
}
\value{
A single numeric value with the Normalized Root Mean Squared Error.
}
\description{
Given the observed and predicted values of numeric data computes the
Normalized Root Mean Squared Error.
}
\details{
Normalized Root Mean Squared Error is computed as:\preformatted{![](nrmse.png "sqrt((1 / N) * sum((observed - predicted)^2)) / norm")
}

where y_i is the observed value of element i, \hat{y_i} is the predicted
value of element i, N is the total number of elements and Y' is the
normalized observed values. You can specify one of the following types of
normalization with the \code{type} parameter:
\itemize{
\item \code{"sd"}: Standard deviation.
\item \code{"mean"}: Mean.
\item \code{"maxmin"} or \code{"range"}: Maximun minus minimum (range).
\item \code{"iqr"}: Interquantile range (Q3 - Q1).
}
}
\examples{
\dontrun{
set.seed(1)
x <- rnorm(100)
nrmse(x, x)
nrmse(x, x - 1, type = "mean")
nrmse(x, x + 10, type = "iqr")
nrmse(x, x + 10, type = "range")
nrmse(x, x + 10, type = "maxmin")
}

}
\seealso{
Other numeric_metrics: 
\code{\link{maape}()},
\code{\link{mae}()},
\code{\link{mse}()},
\code{\link{numeric_summary}()},
\code{\link{pearson}()},
\code{\link{rmse}()},
\code{\link{spearman}()}
}
\concept{numeric_metrics}
