% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{source_zip_list}
\alias{source_zip_list}
\title{source_zip_list}
\format{a character vector of length 51 includes one 5 digits zip code selected from
 each state, where the most central zip code in each state selected.}
\usage{
source_zip_list
}
\description{
a list of zip code used in skm package demonstration.
}
\keyword{datasets}

